requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pymongo==4.6.1
python-dotenv==1.0.0
typing_extensions==4.9.0
Werkzeug==3.0.1
